+++ pytest/tests/__init__.py
@@ -3,12 +3,7 @@
 import os
 import shutil
 from pathlib import Path
-
-import para
-
-para.logging.init_rich_console()
-prev_input = para.logging.get_rich_console().input
-para.logging.set_avoid_print_banner_overwrite(True)
+import logging
 
 
 def resolve_test_path() -> Path:
@@ -35,17 +30,6 @@ def resolve_test_path() -> Path:
 
 
 BASE_TEST_PATH = resolve_test_path()
-
-
-def overwrite_builtin_input(overwrite: str) -> None:
-    """ Overwrites the input with a lambda that returns the specified value """
-    getattr(para.logging, 'output_console').input = \
-        lambda *args, **kwargs: overwrite
-
-
-def reset_input() -> None:
-    """ Resets the output method of the console object """
-    getattr(para.logging, 'output_console').input = prev_input
 
 
 def add_folder(folder_name: str) -> Path: