+++ pytest/tests/compiler/test_process.py
@@ -1,30 +1,21 @@
 # coding=utf-8
 """ Test for the compiler process setup """
+import logging
 from pathlib import Path
 
 from para import initialise_default_paths
 from para.compiler import ProgramCompilationProcess
-from para.logging import set_avoid_print_banner_overwrite
-from .. import add_folder, reset_input, BASE_TEST_PATH
+from .. import add_folder, BASE_TEST_PATH
 
-main_file_path: Path = Path(BASE_TEST_PATH) / "test_files" / "basic.para"
-
-# Avoiding printing the banner (CLI)
-set_avoid_print_banner_overwrite(True)
+logger = logging.getLogger('para')
+logger.setLevel(logging.DEBUG)
+main_file_path: Path = Path(BASE_TEST_PATH) / "test_files" / "main.para"
 
 # Initialises the default paths for the compiler using the work directory
 initialise_default_paths(BASE_TEST_PATH)
 
 
 class TestProcess:
-    @staticmethod
-    def teardown_method(_):
-        """
-        This method is being called after each test case, and it will revert
-        input back to the original function
-         """
-        reset_input()
-
     def test_init(self):
         b_path: Path = add_folder("build")
         d_path: Path = add_folder("dist")