+++ pytest/tests/compiler/test_syntax_check.py
@@ -6,46 +6,34 @@ import os
 from pathlib import Path
 from typing import List
 
-from para import RUNTIME_COMPILER, initialise_default_paths
+from para import initialise_default_paths
 from para.compiler import ParaCompiler, BasicProcess
-from para.logging import set_avoid_print_banner_overwrite
-from .. import reset_input, BASE_TEST_PATH
+from .. import BASE_TEST_PATH
 
-logger = logging.getLogger('compiler')
-
+logger = logging.getLogger('para')
+logger.setLevel(logging.DEBUG)
 compiler = ParaCompiler()
 
-# Initialises the logging session for the compiler
-RUNTIME_COMPILER.init_logging_session(
-    level=logging.DEBUG, print_banner=False
-)
-
-main_file_path: Path = BASE_TEST_PATH / "test_files" / "basic.para"
+main_file_path: Path = BASE_TEST_PATH / "test_files" / "main.para"
 test_c_files_dir: Path = BASE_TEST_PATH / "test_files" / "c_ref_files"
 test_para_files_dir: Path = BASE_TEST_PATH / "test_files"
-
-# Avoiding printing the banner (CLI)
-set_avoid_print_banner_overwrite(True)
 
 # Initialises the default paths for the compiler using the work directory
 initialise_default_paths(BASE_TEST_PATH)
 
 
-class TestSyntaxCheck:
-    @staticmethod
-    def teardown_method(_):
-        """
-        This method is being called after each test case, and it will revert
-        input back to the original function
-        """
-        reset_input()
-
-    def test_single_file(self):
+class TestValidateSyntax:
+    def test_single_file_without_prefer_logging(self):
         asyncio.run(
-            RUNTIME_COMPILER.validate_syntax(main_file_path, 'utf-8', False)
+            compiler.validate_syntax(main_file_path, 'utf-8', False)
         )
 
-    def test_para_files(self):
+    def test_single_file_with_prefer_logging(self):
+        asyncio.run(
+            compiler.validate_syntax(main_file_path, 'utf-8', True)
+        )
+
+    def test_multiple_para_files(self):
         files: List[os.DirEntry] = []
 
         for entry in os.scandir(test_para_files_dir):
@@ -55,12 +43,10 @@ class TestSyntaxCheck:
 
         for file in files:
             asyncio.run(
-                RUNTIME_COMPILER.validate_syntax(
-                    str(file.path), 'utf-8', False
-                )
+                compiler.validate_syntax(str(file.path), 'utf-8', False)
             )
 
-    def test_c_files(self):
+    def test_multiple_c_files(self):
         files: List[os.DirEntry] = []
 
         for entry in os.scandir(test_c_files_dir):
@@ -70,7 +56,5 @@ class TestSyntaxCheck:
 
         for file in files:
             asyncio.run(
-                RUNTIME_COMPILER.validate_syntax(
-                    str(file.path), 'utf-8', False
-                )
+                compiler.validate_syntax(str(file.path), 'utf-8', False)
             )